# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: HackWerk HTTP API
  version: 0.4.0
  description: Interner Vertrag des serverseitig gerenderten HackWerk-Webdienstes.
servers:
  - url: http://localhost:18533
paths:
...
  /customers:
    get:
      operationId: listCustomers
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Serverseitig paginierte aktive Kundenliste ohne Kontaktwerte in der URL. }
        "401": { description: Anmeldung erforderlich. }
    post:
      operationId: createCustomerIntake
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { $ref: "#/components/schemas/CustomerIntakeForm" }
      responses:
        "303": { description: Kunde, Auftrag und Wartelisteneintrag atomar angelegt; Redirect zur Kundenakte. }
        "403": { description: Berechtigung oder CSRF fehlt. }
        "422": { description: Kontakt-, Mengen-, Dauer-, Zeitraum- oder Transportvalidierung fehlgeschlagen. }
  /customers/search:
    post:
      operationId: searchCustomers
      description: POST verhindert, dass Telefonnummern oder E-Mail-Adressen in URL, Browserhistorie oder Access Logs gelangen.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [csrf_token]
              properties:
                csrf_token: { type: string, writeOnly: true }
                q: { type: string, maxLength: 320 }
                page: { type: integer, minimum: 1, default: 1 }
      responses:
        "200": { description: Suchergebnis nach Name, Firma, Ort, Telefon oder Auftragsnummer. }
        "403": { description: Berechtigung oder CSRF fehlt. }
  /customers/{customerID}:
    parameters:
      - $ref: "#/components/parameters/CustomerID"
    get:
      operationId: getCustomer
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Kundenakte mit getrennten Aufträgen, append-only Notizen und sicherem Maps-Link. }
        "404": { description: Kundenakte nicht gefunden. }
    post:
      operationId: updateCustomer
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { $ref: "#/components/schemas/CustomerUpdateForm" }
      responses:
        "303": { description: Kundendaten aktualisiert. }
        "403": { description: Berechtigung oder CSRF fehlt. }
        "409": { description: Erwartete Version ist veraltet. }
        "422": { description: Eingabe fachlich ungültig. }
  /customers/{customerID}/archive:
    parameters:
      - $ref: "#/components/parameters/CustomerID"
    post:
      operationId: archiveCustomer
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { $ref: "#/components/schemas/VersionForm" }
      responses:
        "303": { description: Kunde archiviert; Historie bleibt erhalten. }
        "403": { description: Adminberechtigung oder CSRF fehlt. }
        "409": { description: Erwartete Version ist veraltet. }
  /customers/{customerID}/jobs:
    parameters:
      - $ref: "#/components/parameters/CustomerID"
    post:
      operationId: createCustomerJob
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { $ref: "#/components/schemas/JobForm" }
      responses:
        "303": { description: Separater Auftrag und Wartelisteneintrag für bestehenden aktiven Kunden angelegt. }
        "403": { description: Berechtigung oder CSRF fehlt. }
        "404": { description: Aktiver Kunde nicht gefunden. }
        "422": { description: Auftrag fachlich ungültig. }
  /jobs/{jobID}:
    parameters:
      - $ref: "#/components/parameters/JobID"
    post:
      operationId: updateJob
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              allOf:
                - $ref: "#/components/schemas/JobForm"
                - $ref: "#/components/schemas/VersionForm"
      responses:
        "303": { description: Auftrag aktualisiert. }
        "403": { description: Berechtigung oder CSRF fehlt. }
        "409": { description: Erwartete Version ist veraltet oder Auftrag nicht mehr bearbeitbar. }
        "422": { description: Auftrag fachlich ungültig. }
  /jobs/{jobID}/archive:
    parameters:
      - $ref: "#/components/parameters/JobID"
    post:
      operationId: archiveJob
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { $ref: "#/components/schemas/VersionForm" }
      responses:
        "303": { description: Auftrag archiviert und aktiver Wartelisteneintrag atomar entfernt. }
        "403": { description: Adminberechtigung oder CSRF fehlt. }
        "409": { description: Erwartete Version ist veraltet oder Auftrag nicht archivierungsfähig. }
  /jobs/{jobID}/notes:
    parameters:
      - $ref: "#/components/parameters/JobID"
    post:
      operationId: appendJobNote
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [csrf_token, body]
              properties:
                csrf_token: { type: string, writeOnly: true }
                body: { type: string, minLength: 1, maxLength: 4000 }
                customer_id: { type: string, format: uuid }
      responses:
        "303": { description: Neue unveränderliche Notiz mit Autor und Zeitstempel angehängt. }
        "403": { description: Berechtigung oder CSRF fehlt. }
        "422": { description: Notiz leer oder zu lang. }
  /waitlist:
    get:
      operationId: listWaitlist
      security: [{ sessionCookie: [] }]
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string, enum: [chipping_only, chipping_with_transport] } }
        - { name: region, in: query, schema: { type: string } }
        - { name: urgency, in: query, schema: { type: string, enum: [low, normal, high, urgent] } }
        - { name: month, in: query, schema: { type: string, pattern: "^[0-9]{4}-(0[1-9]|1[0-2])$" } }
        - { name: sort, in: query, schema: { type: string, enum: [entered, preferred, urgency, volume, region], default: entered } }
        - { name: direction, in: query, schema: { type: string, enum: [asc, desc], default: asc } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
      responses:
        "200": { description: Responsive, stabil paginierte und serverseitig sortierte/gefilterte Warteliste. }
        "401": { description: Anmeldung erforderlich. }
  /waitlist/{waitlistID}/priority:
    parameters:
      - $ref: "#/components/parameters/WaitlistID"
    post:
      operationId: updateWaitlistPriority
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [csrf_token, version, priority]
              properties:
                csrf_token: { type: string, writeOnly: true }
                version: { type: integer, minimum: 1 }
                priority: { type: integer, minimum: -100, maximum: 100 }
      responses:
        "303": { description: Adminpriorität aktualisiert und auditiert. }
        "403": { description: Adminberechtigung oder CSRF fehlt. }
        "409": { description: Erwartete Version ist veraltet. }
  /waitlist/{waitlistID}/remove:
    parameters:
      - $ref: "#/components/parameters/WaitlistID"
    post:
      operationId: removeWaitlistEntry
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [csrf_token, version, reason]
              properties:
                csrf_token: { type: string, writeOnly: true }
                version: { type: integer, minimum: 1 }
                reason: { type: string, enum: [scheduled, cancelled, duplicate, other] }
      responses:
        "303": { description: Wartelisteneintrag historisch entfernt und auditiert. }
        "403": { description: Adminberechtigung oder CSRF fehlt. }
        "409": { description: Erwartete Version ist veraltet. }
...
components:
  parameters:
    CustomerID:
      name: customerID
      in: path
      required: true
      schema: { type: string, format: uuid }
    JobID:
      name: jobID
      in: path
      required: true
      schema: { type: string, format: uuid }
    WaitlistID:
      name: waitlistID
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    VersionForm:
      type: object
      required: [csrf_token, version]
      properties:
        csrf_token: { type: string, writeOnly: true }
        version: { type: integer, minimum: 1 }
        customer_id: { type: string, format: uuid }
    CustomerFields:
      type: object
      required: [notification]
      properties:
        first_name: { type: string, maxLength: 200 }
        last_name: { type: string, maxLength: 200 }
        company_name: { type: string, maxLength: 200 }
        street: { type: string, maxLength: 250 }
        postal_code: { type: string, maxLength: 32 }
        locality: { type: string, maxLength: 120 }
        region: { type: string, maxLength: 120 }
        address_freeform: { type: string, maxLength: 1000 }
        phone: { type: string, maxLength: 64 }
        email: { type: string, format: email, maxLength: 320 }
        notification: { type: string, enum: [email, sms, both, none] }
    CustomerUpdateForm:
      allOf:
        - $ref: "#/components/schemas/CustomerFields"
        - $ref: "#/components/schemas/VersionForm"
    JobForm:
      type: object
      required: [csrf_token, job_type, volume_m3, hack_duration, transport_mode, urgency, source]
      properties:
        csrf_token: { type: string, writeOnly: true }
        job_type: { type: string, enum: [chipping_only, chipping_with_transport] }
        volume_m3: { type: string, pattern: "^[0-9]+([.,][0-9]+)?$" }
        hack_duration: { type: string, description: Minuten oder Stunden:Minuten. }
        transport_duration: { type: string, description: Minuten oder Stunden:Minuten. }
        transport_trips: { type: integer, minimum: 0 }
        transport_mode: { type: string, enum: [none, internal, external, undecided] }
        external_confirmed: { type: boolean, default: false }
        preferred_start: { type: string, format: date }
        preferred_end: { type: string, format: date }
        preference_text: { type: string, maxLength: 1000 }
        urgency: { type: string, enum: [low, normal, high, urgent] }
        region: { type: string, maxLength: 120 }
        source: { type: string, enum: [phone, voice, email, in_person, other] }
        note: { type: string, maxLength: 4000 }
    CustomerIntakeForm:
      allOf:
        - $ref: "#/components/schemas/CustomerFields"
        - $ref: "#/components/schemas/JobForm"
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hackplan_session